feat: add render function to sponsor logo and enhance scroll collapse handling; refactor header stories for improved structure
