+++ packages/app-sundevil/src/components/Header/stories/Content.js
@@ -1,0 +1,15 @@
+import styled from "styled-components";
+
+export const Content = styled.div`
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  height: 100vh;
+  background: repeating-linear-gradient(
+    45deg,
+    #efefef,
+    #dfdfdf 10px,
+    #efefef 10px,
+    #efefef 20px
+  );
+`;
