+++ packages/component-header/src/core/models/types.js
@@ -16,6 +16,7 @@
  * @property {string} [adId]
  * @property {function} [onClick]
  * @property {function} [onFocus]
+ * @property {function} [render]
  */
 
 /**
